requirements.txt
replicate
python-dotenv
streamlit
